requirements.txt
# Python dependencies for WakaTime automation
# Pin versions for security and reproducibility

# HTTP requests library
requests==2.31.0

# Google Generative AI SDK
google-generativeai==0.3.2
